requirements.txt
Django==1.7.4
django-tastypie==0.12.1
MySQL-python==1.2.5
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
